Refactor: split navigation and hero section
I wanted a fixed navigation at the top, to make jumping between sections easy.
The navigation and hero section was split as I wanted a black background to the
navigation component but keep the hero section the default colour (white).

Needed to add a class to the body tag of the index.html page in order that
page content would not start underneath the navbar, otherwise the hero would
always be partially burried under the navbar and links to different sections
would be partially covered by the navbar.