Added current sponsor section
Updated the app-state to hold details of the current sponsor.  So when we run
another event then we can just updated the app-state with new sponsor details
and it will update the page content.

The call to `content/sponsor-current` was given the specific part of the
`app-state` data as an argument.

I wanted to make the sponsor section a little narrower, as its a relatively
smaller section and making it narrower would help to focus on the message.  This
should make the sponsor stand out, but without detracting from the other content
too much.  `is-8` and `is-offset-2` modifiers were used to create the desired
effect.

`has-text-centered` modifier was used to center the title of this section