Updated Install section - new image
Created an image using the logo's of the 4 development tools we recommend at
ClojureBridge.

Logos were copied into a Google slides doc and then a screen grab was created
and saved as a pdf.  I normally do this in Inkscape, but the Coach training
presentation already had some of the logos...