Change from hero to columns for top banner
Removed the hero style for the top banner and changed to a simple columns
approach. This makes more use of the page space and registration buttons are
easy to find (and responsive).

As we dont have a large ClojureBridge image for the hero style, it doesnt make
sense to use that style.