Separate content sections with a level class and size 5 title
Using a container provides no padding between content sections (which seems to
work okay as we use box for padding around the content.

However, the navbar links jump to the start of the component, positioning that
at the top of the page rather than after the end of the navbar.  Ideally there
should be a fix for that, but am unable to find anything specific.

Instead forcing a separator between the content of the right size to prevent a
link jump to a section from drawing the content underneath the fixed navbar.

Its a bit of a hack, but it does work and gives a nice gap between the content.